Add hello world unsigned char to int/float functions
Define functions unsignedCharToInt and unsignedChartoFloat functions
that simply return 0. In this version, the receivePacket() does nothing more
than define an integer, packetType, which gets set to 0. It then checks the
value of packetType and will assign the member Processor::packetType to "Audio".

There is no actual packet processing in this version.