Implement unsignedCharToInt and add test
Implement unsignedCharToInt assuming a little endian system. This implementation does not
use a for-loop, but brute force assumes that an integer is four bytes long. Add a test
to check that, as a first pass, this function is doing what it should.

Include unsignedCharToInt, unsignedCharToFloat, and isLittleEndian to the processor.h header file.
My system is little-endian.