Add test for getPacketType
The function processPacket accepts the test packet as a series of
unsigned chars. With this modification, the test inputs had to be modified.
For example, an audio packet will have the first four bytes define an integer
of zero. As such, testPacket[0], ..., testPacket[4] must all be 0x00.

I've also added, in comment, a prototype and a possible implementation for
an intToUnsignedChar function.